requirements.txt
fastapi
uvicorn
pandas
python-dotenv
pydantic 